ols: move sigrok channel creation from protocol.c to api.c
It's unexpected that several code paths redundantly create sigrok
channels, and that most of them reside in the protocol.c source file.
Concentrate channel creation in api.c instead in the scan() routine,
to remain aware during maintenance.

Fallback to the maximum channel count for SUMP compatibles, prefer
metadata gathered from the device when available, including quirks for
models which are known to deviate. Just setup the number during probe,
and create the channels at the end of scan() in common code.

Change the channel count struct member to unsigned while we are here.
Shorten a long comment line to not have to break it.

+++ src/hardware/openbench-logic-sniffer/protocol.h
@@ -100,7 +100,7 @@
 
 struct dev_context {
 	/* constant device properties: */
-	int max_channels;
+	size_t max_channels;
 	uint32_t max_samples;
 	uint32_t max_samplerate;
 	uint32_t protocol_version;
